requirements.txt
streamlit
pdf2image
pytesseract
pillow
pandas
opencv-python
google-generativeai
python-dotenv
